Changed backend behind the waves
Instead of giving the GPU full control of the wave handling, we now take a big chunk of that responsibility back to the CPU.
The CPU is now responsible for all the wave data, increasing the size etc..
The GPU is only responsible for during this data into a height map. Even with high quality texture and multiple wave performance remains pretty consistent.